requirements.txt
requests
psycopg2-binary
pandas
streamlit
cloudscraper
beautifulsoup4